imgtool: Provide for preloading a DICE TLV
Layer 0 of the Device Identifier Composition Engine (DICE) can optionally be
equipped with a Cert_L0, a certificate for the public portion of the
DeviceID. Doing so obviates distributing the public portion of the DeviceID
to communication partners. However, as the DeviceID varies with the hash of
the Layer 0 software, Cert_L0 cannot simply be put in the Layer 0 software.
This would change the DeviceID and hence invalidate Cert_L0. As a solution,
this commit introduces an unprotected TLV for passing a Cert_L0 to Layer 0.

When not preloading Cert_L0, this newly introduced "DICE TLV" can
alternatively serve to configure the subject field of Cert_L1. In that case,
the DICE TLV must be placed in the protected TLV area so as to affect the
DeviceID. Note that setting the subject field is unnecessary when preloading
Cert_L0 as Cert_L1 then inherits Cert_L0's subject field automatically.