scripts: Add script for issuing Cert_L0 and s_L0
DICE is being configured by attaching a TLV to the Layer 0 image. In
particular, one can equip Layer 0 with a Cert_L0, which is akin to an
IEEE IDevID certificate. This commit adds a script for issuing such a
Cert_L0 and producing the contents of the DICE TLV. Note that, in the case
of TinyDICE, Cert_L0 is an implicit certificate and hence the CA issues
not only Cert_L0 but also private key reconstruction data s_L0 along with
Cert_L0. For further information on implicit certificates, see the SEC4
standard by the Standards for Efficient Cryptography Group (SECG).