bootutil: Add support for the Device Identifier Composition Engine (DICE)
Two main features of DICE are to enable remote attestation at runtime and
to keep long-term secrets in a trusted first-stage bootloader, thereby
obviating rekeying when vulnerabilities in upper layers show up, such as in
the operating system. This commit adds suport for Siemens' lightweight
version of DICE dubbed TinyDICE. However, based on this, it should be
straightforward to add support for other versions of DICE, such as Google's
Open Profile for DICE.

The general concept is to run two cascaded instances of MCUboot, namely a
root of trust (RoT) and Layer 0 in DICE parlance. The RoT should be
compiled with minimal functionality and only serves to derive a symmetric
key CDI_L0 for the second instance of MCUboot. That instance, on the other
hand, does not contribute to the trusted code and can therefore be
compiled with advanced features of MCUboot.