bootutil/crypto: Add general-purpose HMAC and HKDF implementations
Using the currently configured crypto back end, this implements HMAC and
HKDF. This implementation serves as a fallback solution for crypto back
ends that do not provide these functions. As there are plans to remove
TinyCrypt from MCUboot, no effort is made to use TinyCrypt when
MCUBOOT_USE_TINYCRYPT is set. Instead, this configuration falls back on
the general-purpose HMAC and HKDF implementations.

Signed-off-by: kkrentz <konrad.krentz@gmail.com>