bootutil/crypto: Silently abort a series of calls to bootutil_sha_*
The current API actually requires checking each call to bootutil_sha_init,
bootutil_sha_update, bootutil_sha_finish, and bootutil_sha_abort for errors.
This would result in a lot of clutter and is not done at the moment. This
commit wraps these functions so as to silently ignore subsequent calls in
a series upon errors. As a result, only bootutil_sha_finish needs error
checking.